Refactor Course model and introduce media handling
In this commit, the Course model has been refactored to use the HasFactory, SoftDeletes, and InteractsWithMedia traits. The model also now implements the HasMedia interface, enabling the manipulation of media files related to course instances.  A registerMediaCollections function was added and the model's casting has been updated for better date formatting.