Refine article resource transformation logic
This update revises the `ArticleJsonCollection` transformation logic to handle grouping by type only when no type is specified in the route. It also improves handling of `take` requests to limit the number of articles per type. In addition, the date format used in `ArticleJsonResource` has been simplified to use the raw, unformatted date-time data.