Implement WithEnum to optimize and streamline QueryBuilder usage
A new enumerated class, WithEnum, was added to hold predefined strings for optimizing QueryBuilder's ->with() method. This not only helps in reducing the data sent to the client by fetching only essential attributes, but it also improves code readability and maintainability by minimizing hard-coded strings. Changes were made on the Article and Course controllers to adapt to this implementation.