Refactor Course Controller for better autocomplete functionality.
Modified the CourseController to pass a query instead of a model to the Spatie QueryBuilder, which enables better autocomplete functionality. This change ensures that only published courses are fetched, by adding a 'published' method call in the query passed to QueryBuilder. It also enhances the readability and maintainability of the code by improving the structure of chained method calls.