Update article and course resources, adjust migrations and seeders
The modifications primarily aim to adjust the Filament resources for Articles and Courses by extending their functionalities and refining the appearance of the forms. In terms of database configuration, a cascade delete option for 'course_id' within the articles_courses pivot table was implemented. Furthermore, during seeding data, the creation of a User factory class was made universally accessible instead of limiting it to local environments.